+++ app/src/main/java/reikai/presentation/details/MangaEntryAdapter.kt
@@ -17,6 +17,7 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import reikai.domain.entry.EntryId
 import reikai.domain.merge.ChapterGap
+import reikai.presentation.components.chapterSubtitle
 import tachiyomi.domain.chapter.model.Chapter
 import tachiyomi.domain.library.service.LibraryPreferences
 import tachiyomi.domain.manga.model.Manga